Modularize cloud web workbench
